backend/requirements.txt
# FastAPI and ASGI server
fastapi==0.109.0
uvicorn[standard]==0.27.0

# Database
sqlmodel==0.0.14
psycopg2-binary==2.9.9

# Authentication and Security
python-jose[cryptography]==3.3.0
python-multipart==0.0.6
passlib[bcrypt]==1.7.4

# Environment and Configuration
python-dotenv==1.0.0

# CORS and middleware
fastapi-cors==0.0.6

# Validation
pydantic==2.5.3
pydantic-settings==2.1.0
